refactor(home): extract sections to components
